все работает, маленькие правки остались
