+++ Task3/Task3/Image.swift
@@ -9,7 +9,6 @@ import UIKit
 
 enum Images: String {
     case test
-
     var image: UIImage {
         guard let image = UIImage(named: self.rawValue) else { assertionFailure(); return UIImage() }
         return image
