+++ Task3/Task3/SceneDelegate.swift
@@ -9,18 +9,18 @@ import UIKit
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
-    var window: UIWindow?
+	var window: UIWindow?
+	@available(iOS 13.0, *)
+	func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+		guard let windowScene = scene as? UIWindowScene else { return assertionFailure() }
+		window = UIWindow(frame: windowScene.coordinateSpace.bounds)
 
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        guard let windowScene = scene as? UIWindowScene else { return assertionFailure() }
-        window = UIWindow(frame: windowScene.coordinateSpace.bounds)
+		guard let window = window else { return assertionFailure() }
+		window.windowScene = windowScene
 
-        guard let window = window else { return assertionFailure() }
-        window.windowScene = windowScene
-
-        let viewController = TabBarController()
-        let navigationController = UINavigationController(rootViewController: viewController)
-        window.rootViewController = navigationController
-        window.makeKeyAndVisible()
-    }
+		let viewController = TabBarController()
+		let navigationController = UINavigationController(rootViewController: viewController)
+		window.rootViewController = navigationController
+		window.makeKeyAndVisible()
+	}
 }
